rename compute-kzg-proof to compute-proof to include view proof generation
