+++ .gitignore
@@ -2,3 +2,5 @@ target/
 data/
 optimism/
 tags
+rollup.json
+resources/
