+++ Cargo.toml
@@ -19,7 +19,7 @@ hokulea-client-bin = { path = "bin/client", version = "0.1.0", default-features 
 hokulea-client = { path = "crates/client", version = "0.1.0", default-features = false }
 hokulea-eigenda = { path = "crates/eigenda", version = "0.1.0", default-features = false }
 hokulea-proof = { path = "crates/proof", version = "0.1.0", default-features = false }
-hokulea-compute-kzg-proof = { path = "crates/compute-kzg-proof", version = "0.1.0", default-features = false }
+hokulea-compute-proof = { path = "crates/compute-proof", version = "0.1.0", default-features = false }
 
 # Maili
 maili-consensus = { version = "0.2.8", default-features = false, features = ["serde"] }
