+++ crates/client/Cargo.toml
@@ -21,5 +21,5 @@ tracing.workspace = true
 async-trait.workspace = true
 
 rust-kzg-bn254-primitives.workspace = true
-hokulea-compute-kzg-proof.workspace = true
+hokulea-compute-proof.workspace = true
 alloy-primitives.workspace = true
