+++ crates/client/src/lib.rs
@@ -1,5 +1,3 @@
 pub mod core_client;
 pub mod witgen_client;
 pub mod witness;
-
-// TODO we might need to decouple the library a bit more to separate STD, NO_STD
