+++ crates/compute-proof/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hokulea-compute-kzg-proof"
+name = "hokulea-compute-proof"
 version = "0.1.0"
 edition = "2021"
 
@@ -10,3 +10,4 @@ rust-kzg-bn254-primitives.workspace = true
 num.workspace = true
 
 alloy-primitives.workspace = true
+risc0-zkvm.workspace = true
