+++ crates/compute-proof/src/lib.rs
@@ -9,6 +9,6 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
-pub mod compute_kzg_proof;
+pub mod kzg_proof;
 
-pub use compute_kzg_proof::compute_kzg_proof;
+pub use kzg_proof::compute_kzg_proof;
