+++ crates/compute-proof/src/view_proof.rs
@@ -1,0 +1,12 @@
+use risc0_zkvm::Receipt;
+
+/// This function computes view proof. TODO abstract away Reciept to make it 
+/// generic to multiple zkVM
+fn compute_view_proof() -> Receipt {
+    compute_view_proof_steel()
+}
+
+/// compute view proof using Steel
+fn compute_view_proof_steel() -> Receipt {
+    unimplemented!()
+}
